fix: moved toBest tests in correct file
